+++ app/components/listings/ListingReservation.tsx
@@ -14,6 +14,7 @@ interface ListingReservationProps {
   onSubmit: () => void;
   disabled?: boolean;
   disabledDates: Date[];
+  ownerIsWatching: boolean;
 }
 
 const ListingReservation: React.FC<ListingReservationProps> = ({
@@ -24,25 +25,14 @@ const ListingReservation: React.FC<ListingReservationProps> = ({
   onSubmit,
   disabled,
   disabledDates,
+  ownerIsWatching,
 }) => {
   const { t } = useTranslation();
 
   return (
     <div className="flex flex-col gap-8">
-      <div
-        className="
-      bg-white 
-        rounded-xl 
-        border-[1px]
-      border-neutral-200 
-        overflow-hidden
-        shadow-xl
-      "
-      >
-        <div
-          className="
-      flex flex-row items-center gap-1 p-4"
-        >
+      <div className="bg-white rounded-xl border-[1px]border-neutral-200 overflow-hidden shadow-xl">
+        <div className="flex flex-row items-center gap-1 p-4">
           <div className="text-2xl font-semibold">
             {price.toFixed(2)}{" "}
             <span className="font-light text-lg text-neutral-600">
@@ -56,25 +46,22 @@ const ListingReservation: React.FC<ListingReservationProps> = ({
             value={dateRange}
             disabledDates={disabledDates}
             onChange={(value) => onChangeDate(value.selection)}
+            direction="vertical"
           />
         </div>
         <hr />
         <div className="p-4">
-          <Button disabled={disabled} label={t("Reserve")} onClick={onSubmit} />
+          <Button
+            disabled={disabled}
+            label={ownerIsWatching ? t("Mark_as_unavailable") : t("Reserve")}
+            onClick={onSubmit}
+          />
         </div>
         <hr />
-        <div
-          className="
-          p-4 
-          flex 
-          flex-row 
-          items-center 
-          justify-between
-          font-semibold
-          text-lg
-        "
-        >
-          <div>{t("Total_price_of_reservation")}</div>
+        <div className="p-4 flex flex-row items-center justify-between font-semibold text-lg">
+          <div className="font-light text-sm">
+            {t("Total_price_of_reservation")}
+          </div>
           <div>
             {totalPrice.toFixed(2)} {t("BGN")}
           </div>